pyproject.toml
[project]
name = "headline-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-utils-traveler (>=0.0.11,<0.0.12)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
